pyproject.toml
[tool.poetry]
name = "scripts"
version = "0.1.0"
description = ""
authors = ["Wojciech Malarski <w.malarski@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
gql = {version = "3.0.0a5", extras = ["all"]}
python-dotenv = "^0.15.0"
tqdm = "^4.55.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
